# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VyOS API",
    "description": "REST proxy for one or more VyOS router devices. Exposes CRUD operations over interfaces, VRFs, VLANs, firewall policies, and firewall address groups. Each device is identified by the name supplied in the VYOS_HOSTS environment variable.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/devices/{device_id}/firewall/address-groups": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" }
      ],
      "get": {
        "tags": ["address-groups"],
        "summary": "List all address groups",
        "operationId": "listAddressGroups",
        "responses": {
          "200": {
            "description": "Address group list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AddressGroupInfo" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "post": {
        "tags": ["address-groups"],
        "summary": "Create an address group",
        "description": "Creates a new address group and populates it with the supplied addresses. An empty `addresses` list creates an empty group.",
        "operationId": "createAddressGroup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAddressGroupRequest" },
              "example": {
                "name": "RFC1918",
                "addresses": ["10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"],
                "description": "private-ranges"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address group created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddressGroupInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    },
    "/devices/{device_id}/firewall/address-groups/{group}": {
      "parameters": [
        { "$ref": "#/components/parameters/device_id" },
        { "$ref": "#/components/parameters/group" }
      ],
      "get": {
        "tags": ["address-groups"],
        "summary": "Get an address group",
        "operationId": "getAddressGroup",
        "responses": {
          "200": {
            "description": "Address group details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddressGroupInfo" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "put": {
        "tags": ["address-groups"],
        "summary": "Replace an address group",
        "description": "Full replacement: deletes all existing members, then adds the supplied addresses. An empty list clears the group.",
        "operationId": "updateAddressGroup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAddressGroupRequest" },
              "example": {
                "addresses": ["192.168.0.0/24", "192.168.1.0/24"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address group replaced",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddressGroupInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      },
      "delete": {
        "tags": ["address-groups"],
        "summary": "Delete an address group",
        "operationId": "deleteAddressGroup",
        "responses": {
          "204": { "description": "Address group deleted" },
          "404": { "$ref": "#/components/responses/DeviceNotFound" },
          "422": { "$ref": "#/components/responses/DeviceRejected" },
          "502": { "$ref": "#/components/responses/DeviceError" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "device_id": {
        "name": "device_id",
        "in": "path",
        "required": true,
        "description": "Device name as registered in VYOS_HOSTS",
        "schema": { "type": "string", "example": "router1" }
      },
...
      "group": {
        "name": "group",
        "in": "path",
        "required": true,
        "description": "Address group name",
        "schema": { "type": "string", "example": "RFC1918" }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "device not found: router99" }
        }
      },
...
      "AddressGroupInfo": {
        "type": "object",
        "required": ["name", "addresses"],
        "properties": {
          "name":        { "type": "string", "example": "RFC1918" },
          "addresses":   { "type": "array", "items": { "type": "string" }, "example": ["10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"] },
          "description": { "type": "string", "example": "private-ranges" }
        }
      },
      "CreateAddressGroupRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name":        { "type": "string", "example": "RFC1918" },
          "addresses":   { "type": "array", "items": { "type": "string" }, "description": "Initial members. Empty list creates an empty group.", "example": ["10.0.0.0/8"] },
          "description": { "type": "string", "example": "private-ranges" }
        }
      },
      "UpdateAddressGroupRequest": {
        "type": "object",
        "required": ["addresses"],
        "properties": {
          "addresses":   { "type": "array", "items": { "type": "string" }, "description": "Complete replacement address list", "example": ["192.168.0.0/24"] },
          "description": { "type": "string", "example": "updated-description" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Missing or invalid request fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface, type, and address are required" }
          }
        }
      },
      "NotFound": {
        "description": "Device or resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "interface not found" }
          }
        }
      },
      "DeviceNotFound": {
        "description": "The requested device_id is not registered",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device not found: router99" }
          }
        }
      },
      "DeviceRejected": {
        "description": "The VyOS device rejected the operation (invalid config, constraint violation, etc.)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device rejected operation: Configuration path already exists" }
          }
        }
      },
      "DeviceError": {
        "description": "Could not communicate with the VyOS device (network error, TLS failure, timeout)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "device communication error: dial tcp 192.168.1.1:443: connection refused" }
          }
        }
      }
    }
  }
}